backend/requirements.txt
asgiref==3.7.2
black==23.12.1
boto3==1.34.23
botocore==1.34.23
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
smsir-python==1.0.7
sqlparse==0.4.4
urllib3==2.0.7
django-filter==24.2